requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.2
async-timeout==4.0.3
attrs==25.1.0
Babel==2.9.1
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.1
python-decouple==3.8
pytz==2025.1
yarl==1.18.3